+++ packages/router/router.ts
@@ -1225,10 +1225,13 @@ export function createRouter(init: RouterInit): Router {
       return;
     }
 
-    // Short circuit if it's only a hash change and not a mutation submission
+    // Short circuit if it's only a hash change and not a mutation submission.
+    // Ignore on initial page loads because since the initial load will always
+    // be "same hash".
     // For example, on /page#hash and submit a <Form method="post"> which will
     // default to a navigation to /page
     if (
+      state.initialized &&
       isHashChangeOnly(state.location, location) &&
       !(opts && opts.submission && isMutationMethod(opts.submission.formMethod))
     ) {
@@ -2474,7 +2477,13 @@ export function createRouter(init: RouterInit): Router {
   }
 
   function _internalSetRoutes(newRoutes: AgnosticDataRouteObject[]) {
-    inFlightDataRoutes = newRoutes;
+    manifest = {};
+    inFlightDataRoutes = convertRoutesToDataRoutes(
+      newRoutes,
+      mapRouteProperties,
+      undefined,
+      manifest
+    );
   }
 
   router = {
@@ -4009,16 +4018,18 @@ function isHashChangeOnly(a: Location, b: Location): boolean {
   }
 
   if (a.hash === "") {
-    // No hash -> hash
+    // /page -> /page#hash
     return b.hash !== "";
   } else if (a.hash === b.hash) {
-    // current hash -> same hash
+    // /page#hash -> /page#hash
     return true;
   } else if (b.hash !== "") {
-    // current hash -> new hash
+    // /page#hash -> /page#other
     return true;
   }
 
+  // If the hash is removed the browser will re-perform a request to the server
+  // /page#hash -> /page
   return false;
 }