+++ packages/react-router-dom/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `react-router-dom`
 
+## 6.11.2
+
+### Patch Changes
+
+- Export `SetURLSearchParams` type ([#10444](https://github.com/remix-run/react-router/pull/10444))
+- Updated dependencies:
+  - `react-router@6.11.2`
+  - `@remix-run/router@1.6.2`
+
 ## 6.11.1
 
 ### Patch Changes